requirements.txt
pyserial>=3.1.1
influxdb>=1.1.0